Changes when notes are active
The original feature request was that notes that were children of notes were inactive on the sheet. This was implemented that all children of notes were inactive on the sheet, which has caused some strife and confusion with users, both new and old.

This new implementation strives to remedy this by disabling notes that are children of other notes, instead of disabling all children of notes.

I still have not set up a local hosting of the project, and unfortunately haven't tested if this works, but from what I can gather reading other parts of the code, it theoretically should, so long as prop.ancestors[-1] looks at the parent of a property (prop.ancestors[0] seems to return the highest ancestor, so I'm assuming [-1] returns the lowest).